+++ src/variables.js
@@ -46,23 +46,6 @@ exports.dlogDetails = dlogDetails;
 var userSettings = { "theme": "auto", "use_custom_theme": false, "theme_background": null, "theme_main": null, "theme_darken_ratio": null, "notification_refresh_interval": 30, "unit": "metric", "radio": "disabled", "radio_type": "tsr", "radio_volume": 100, "display_timezone": Intl.DateTimeFormat().resolvedOptions().timeZone, "data_saver": false, "font_size": "regular", "default_row_per_page": 10 };
 // radio: enabled / disabled / auto-play (enabled)
 // radio-type: tsr / {url}
-if (localStorage.getItem("client-settings") !== null) {
-    try {
-        let lsSettings = JSON.parse(localStorage.getItem("client-settings"));
-        let sKeys = Object.keys(userSettings);
-        for (let i = 0; i < sKeys.length; i++) {
-            if (Object.keys(lsSettings).includes(sKeys[i])) {
-                userSettings[sKeys[i]] = lsSettings[sKeys[i]];
-            }
-        }
-        localStorage.setItem("client-settings", JSON.stringify(userSettings));
-    } catch (error) {
-        console.error("Unable to parse client settings in local storage:");
-        console.error(error);
-    }
-} else {
-    localStorage.setItem("client-settings", JSON.stringify(userSettings));
-}
 exports.userSettings = userSettings;
 
 var userBanner = { name: "Login", role: "", avatar: "https://charlws.com/me.gif" };